Move MatrixFieldNamingStrategy to its own class
