+++ matrix-sdk/src/main/java/org/matrix/androidsdk/util/JsonUtils.java
@@ -20,7 +20,6 @@ package org.matrix.androidsdk.util;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
-import com.google.gson.FieldNamingStrategy;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonArray;
@@ -29,6 +28,7 @@ import com.google.gson.JsonObject;
 
 import org.matrix.androidsdk.data.RoomState;
 import org.matrix.androidsdk.rest.json.ConditionDeserializer;
+import org.matrix.androidsdk.rest.json.MatrixFieldNamingStrategy;
 import org.matrix.androidsdk.rest.model.ContentResponse;
 import org.matrix.androidsdk.rest.model.Event;
 import org.matrix.androidsdk.rest.model.EventContent;
@@ -59,9 +59,7 @@ import org.matrix.androidsdk.rest.model.message.VideoMessage;
 import org.matrix.androidsdk.rest.model.pid.RoomThirdPartyInvite;
 
 import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
-import java.util.Locale;
 import java.util.Map;
 import java.util.TreeSet;
 
@@ -70,42 +68,6 @@ import java.util.TreeSet;
  */
 public class JsonUtils {
     private static final String LOG_TAG = JsonUtils.class.getSimpleName();
-
-    /**
-     * Based on FieldNamingPolicy.LOWER_CASE_WITH_UNDERSCORES.
-     * toLowerCase() is replaced by toLowerCase(Locale.ENGLISH).
-     * In some languages like turkish, toLowerCase does not provide the expected string.
-     * e.g _I is not converted to _i.
-     */
-    public static class MatrixFieldNamingStrategy implements FieldNamingStrategy {
-
-        /**
-         * Converts the field name that uses camel-case define word separation into
-         * separate words that are separated by the provided {@code separatorString}.
-         */
-        private static String separateCamelCase(String name, String separator) {
-            StringBuilder translation = new StringBuilder();
-            for (int i = 0; i < name.length(); i++) {
-                char character = name.charAt(i);
-                if (Character.isUpperCase(character) && translation.length() != 0) {
-                    translation.append(separator);
-                }
-                translation.append(character);
-            }
-            return translation.toString();
-        }
-
-        /**
-         * Translates the field name into its JSON field name representation.
-         *
-         * @param f the field object that we are translating
-         * @return the translated field name.
-         * @since 1.3
-         */
-        public String translateName(Field f) {
-            return separateCamelCase(f.getName(), "_").toLowerCase(Locale.ENGLISH);
-        }
-    }
 
     private static final Gson gson = new GsonBuilder()
             .setFieldNamingStrategy(new MatrixFieldNamingStrategy())
